# services/core/api/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Timeline API
  version: 0.2.0
  description: |
    个人时间线博客 REST API。OpenAPI 是前后端唯一接口合同。
    写入接口使用同源 HttpOnly session cookie，并要求 X-CSRF-Token；创建、提交、导出和上传终结支持 Idempotency-Key。
servers:
  - url: /api/v1
...
paths:
...
  /auth/login/password:
    post:
      tags: [auth]
      operationId: loginPassword
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PasswordLoginRequest'}}}}
      responses:
        '200': {description: 短时 MFA challenge, content: {application/json: {schema: {$ref: '#/components/schemas/MFAChallengeResponse'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /auth/login/totp:
    post:
      tags: [auth]
      operationId: loginTotp
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TOTPLoginRequest'}}}}
      responses:
        '200':
          description: 创建登录 session
          headers:
            Set-Cookie: {$ref: '#/components/headers/SessionCookie'}
          content: {application/json: {schema: {$ref: '#/components/schemas/SessionResponse'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /auth/sessions:
    get:
      tags: [auth]
      operationId: listAuthSessions
      security: [SessionCookie: []]
      responses: {'200': {description: 活动设备, content: {application/json: {schema: {$ref: '#/components/schemas/SessionListResponse'}}}}, '401': {$ref: '#/components/responses/Unauthorized'}}
...
  /auth/recovery/account:
    post:
      tags: [auth]
      operationId: recoverAccount
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RecoveryRequest'}}}}
      responses: {'200': {description: 恢复成功并返回一次性新恢复码与 TOTP setup URI, content: {application/json: {schema: {$ref: '#/components/schemas/RecoveryResponse'}}}}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '429': {description: 尝试次数过多, headers: {Retry-After: {schema: {type: string}}}}}
...
components:
  securitySchemes:
    SessionCookie: {type: apiKey, in: cookie, name: timeline_session, description: HttpOnly SameSite=Lax session cookie; production sets Secure}
...
  headers:
    SessionCookie: {description: HttpOnly session cookie, schema: {type: string}}
  responses:
...
    BadRequest: {description: 请求无效, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    Unauthorized: {description: 未认证, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
...
  schemas:
...
    PasswordLoginRequest: {type: object, required: [password], properties: {password: {type: string, format: password}}}
    MFAChallengeResponse: {type: object, required: [challenge, requiresTotp], properties: {challenge: {type: string}, requiresTotp: {type: boolean}}}
    TOTPLoginRequest: {type: object, required: [challenge, code], properties: {challenge: {type: string}, code: {type: string, minLength: 6, maxLength: 8}}}
    SessionResponse: {type: object, required: [authenticated, csrfToken], properties: {authenticated: {type: boolean}, username: {type: string}, csrfToken: {type: string}, idleExpiresInDays: {type: integer}, absoluteExpiresInDays: {type: integer}}}
    Session: {type: object, required: [id, createdAt, lastSeen], properties: {id: {type: string, format: uuid}, createdAt: {type: string, format: date-time}, lastSeen: {type: string, format: date-time}}}
    SessionListResponse: {type: object, required: [sessions], properties: {sessions: {type: array, items: {$ref: '#/components/schemas/Session'}}}}
    RecoveryRequest: {type: object, required: [recoveryKey, newPassword], properties: {recoveryKey: {type: string}, newPassword: {type: string, format: password, minLength: 12}}}
    RecoveryResponse: {type: object, required: [ok, recoveryKey, totpSetupURI], properties: {ok: {type: boolean}, recoveryKey: {type: string, description: 仅在本次响应显示一次}, totpSetupURI: {type: string, description: 仅在本次响应显示一次}}}
...
    Problem: {type: object, required: [type, title, status, detail], properties: {type: {type: string, format: uri-reference}, title: {type: string}, status: {type: integer}, detail: {type: string}, instance: {type: string}, requestId: {type: string}, fieldErrors: {type: object, additionalProperties: {type: string}}}}
...